requirements.txt
cloudant
python-dotenv
influxdb_client